requirements.txt
requests>=2.32.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dateutil>=2.9.0
tzlocal>=5.2
